Fixing incorrectly nested test class
